feat: Add FindIndexInfoRequset class and enum file
