+++ findex/src/main/java/com/example/findex/dto/indexinfo/CreateIndexInfoRequest.java
@@ -1,16 +1,13 @@
-package com.example.findex.dto.IndexInfo;
+package com.example.findex.dto.indexinfo;
 
-import com.example.findex.entity.SourceType;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record IndexInfoDto(
-    Long id,
+public record CreateIndexInfoRequest(
     String indexClassification,
     String indexName,
-    int employeeItemsCount,
+    int employedItemsCount,
     LocalDate basePointInTime,
     BigDecimal baseIndex,
-    SourceType sourceType,
     boolean favorite) {
 }
