+++ findex/src/main/java/com/example/findex/dto/indexinfo/SortDirectionType.java
@@ -1,0 +1,5 @@
+package com.example.findex.dto.indexinfo;
+
+public enum SortDirectionType {
+  asc, desc
+}
