+++ findex/src/main/java/com/example/findex/dto/indexinfo/SortFieldType.java
@@ -1,0 +1,5 @@
+package com.example.findex.dto.indexinfo;
+
+public enum SortFieldType {
+  indexClassification, indexName, employedItemsCount
+}
